modify 33.cpp

+++ C++/33.cpp
@@ -2,7 +2,7 @@
 
 #include "info.h"
 
-int simpleInterest(int P = 10000, int R = 5, int T = 2) {
+int simpleInterest(float P = 10000, float R = 5, float T = 2) {
   return (P * R * T) / 100;
 }
